feature/DiaryDateCalendar - save date to Redux-state
