--- src/components/DatePicker/index.js
@@ -1,3 +1,0 @@
-import DatePicker from './DatePicker';
-
-export default DatePicker;
